move google vars out of internal
