+++ integrations/google/calendar/watches.go
@@ -7,7 +7,7 @@ import (
 	"go.uber.org/zap"
 	"google.golang.org/api/calendar/v3"
 
-	"go.autokitteh.dev/autokitteh/integrations/google/internal/vars"
+	"go.autokitteh.dev/autokitteh/integrations/google/vars"
 	"go.autokitteh.dev/autokitteh/integrations/internal/extrazap"
 	"go.autokitteh.dev/autokitteh/sdk/sdkservices"
 	"go.autokitteh.dev/autokitteh/sdk/sdktypes"
