+++ integrations/google/gmail/watch.go
@@ -9,7 +9,7 @@ import (
 	"go.uber.org/zap"
 	"google.golang.org/api/gmail/v1"
 
-	"go.autokitteh.dev/autokitteh/integrations/google/internal/vars"
+	"go.autokitteh.dev/autokitteh/integrations/google/vars"
 	"go.autokitteh.dev/autokitteh/integrations/internal/extrazap"
 	"go.autokitteh.dev/autokitteh/sdk/sdkservices"
 	"go.autokitteh.dev/autokitteh/sdk/sdktypes"
